MM-12 채팅 기본 세팅
