+++ src/main/java/com/dev/moim/global/config/socket/WebSocketConfig.java
@@ -1,0 +1,20 @@
+package com.dev.moim.global.config.socket;
+
+import com.dev.moim.global.config.socket.SocketTextHandler;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.socket.config.annotation.EnableWebSocket;
+import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
+import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
+
+@Configuration
+@EnableWebSocket
+public class WebSocketConfig implements WebSocketConfigurer {
+
+
+    @Override
+    public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
+        registry.addHandler(new SocketTextHandler(), "/ws")
+                .setAllowedOrigins("*")
+                .withSockJS();
+    }
+}
